requirements.txt
numpy==1.26.4
openai-whisper
torch==2.2.2
scipy
